requirements.txt
appdirs==1.4.4
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pygubu==0.23.1
pygubu-designer==0.28
pyradios==1.0.2
python-vlc==3.0.16120
requests==2.28.1
screeninfo==0.8
tomli==2.0.1
urllib3==1.26.11
